More than 2x increase large mass speed.

+++ src/extendables/Message/Party.ts
@@ -44,6 +44,7 @@ export async function setupParty(channel: TextChannel, leaderUser: MUser, option
 	const usersWhoConfirmed: string[] = [options.leader.id];
 	let deleted = false;
 	const massTimeout = options.massTimeout ?? Time.Minute * 2;
+	let massStarted = false;
 
 	function getMessageContent(): MessageOptions & MessageEditOptions {
 		const userText =
@@ -133,6 +134,8 @@ export async function setupParty(channel: TextChannel, leaderUser: MUser, option
 			});
 
 			async function startTrip() {
+				if (massStarted) return;
+				massStarted = true;
 				await confirmMessage.delete().catch(noOp);
 				if (!partyCancelled && usersWhoConfirmed.length < options.minSize) {
 					channel.send(`${leaderUser} Not enough people joined your mass!`);
